requirements.txt
unstructured[s3]
unstructured[opensearch]
unstructured[embed-huggingface]
httpx
python-decouple